Slightly Optimize Dance Mooves
This reduces the constant value since sets are not used until the very
end.

Additionally, instead of manually updating every element's set, it now
instead creates one set and gives every element a pointer to that set.
This way, both space and time are saved, since elements in the same
cycle visit the same spots.